Update update_kali.sh
I added some new functions to the system update script to make it more versatile and organized. It now includes a help function that displays usage instructions, making it easier to understand the available options. There is also an argument check to ensure the script is not run without parameters, displaying help if necessary. The operations have been modularized into specific functions, such as updating the package list, upgrading installed packages, performing a distribution upgrade, removing unnecessary packages, and cleaning the package cache. Finally, I used a `case` structure to process the command-line arguments, allowing the user to flexibly and organizedly choose which operations to perform. These improvements make the script easier to use and maintain.